# Performance optimization to Climate RTree search (same as vanilla)