--- a/benchmarks/index.ts
+++ b/benchmarks/index.ts
@@ -1,1 +1,2 @@
 import './math/index.js'
+// import './worldgen/index.js'