--- a/benchmarks/worldgen/index.ts
+++ b/benchmarks/worldgen/index.ts
@@ -1,0 +1,1 @@
+import './MultiNoiseBiomeSource'