Performance optimization in the search of the RTree for the MultiNoiseBiomeSource.

After having tried a different optimization with a Priority Queue, this is the same optimization that exists in the Minecraft code and consists of 2 parts:

1. The currently found closest Leaf is passed to all subsequent calls of search such that the search can be skipped if the node is farther away
2. The tree keeps track of the last result of the `search` call. In the subsequent call this node is passed as the the current closest Leaf to the search of the root node. As subsequent calls are often for areas close to each other this leaf quickly reduces the `dist` to a relatively small number, allowing the skipping of much larger parts of the search.

I've added a benchmark the simulates subsequent calls to similar areas (similar climate target) for the vanilla overworld as well as terralith. For some reason adding the benchmark to the suit makes the `fetch` fail. You can run the benchmark separately with `npx esno benchmarks/worldgen/MultiNoiseBiomeSource.ts`

Results:
|                    | Vanilla                | Terralith             |
|--------------------|------------------------|-----------------------|
| No Optimization    | 170 853 ops/s, ±35.02% | 16 933 ops/s, ±3.81%  |
| Optimization 1     | 207 923 ops/s, ±2.03%  | 67 333 ops/s, ±5.59%  |
| Optimization 1 & 2 | 447 931 ops/s, ±5.04%  | 316 174 ops/s, ±1.44% |

Of course optimization 2 relies on calls from similar areas and as such the performance improvement depends on how similar the calls are. My simulation might not accurately represent the real world generation, but I didn't want to load all the data that would be required to do the proper simulation.