@misode 2023-04-06
Wow this is pretty significant. Through a bit of code digging I discovered this was added in vanilla in 21w37a (first snapshot of 1.18), but not in the experimental snapshots.